Continued to refactor GameRow component
